+++ derive_tests/src/lib.rs
@@ -1,5 +1,5 @@
 use chrono::NaiveDateTime;
-use validify::ValidationError;
+use validify::{schema_err, schema_validation, ValidationError};
 #[allow(unused_imports)]
 use validify::{ValidationErrors, Validify};
 
@@ -74,21 +74,10 @@ fn bar(_a: &str) -> Result<(), ValidationError> {
     Ok(())
 }
 
+#[schema_validation]
 fn validator_test(t: &T) -> Result<(), ValidationErrors> {
-    /* if t.a == "Super no" {
-        field_err!("a", "breh", "Can't be super no", errors);
-    }*/
-    let mut errors = ValidationErrors::new();
     if t.a == "YOLO" && t.e.is_none() {
-        errors.add(
-            ValidationError::new_schema("Invalid YOLO")
-                .with_message("Can't yolo with non existent e".to_string()),
-        );
-    }
-    if !errors.is_empty() {
-        Err(errors)
-    } else {
-        Ok(())
+        schema_err!("Invalid YOLO", "Can't yolo with non existent e");
     }
 }